Some changes to fix the conflict and added the migration.

+++ SM.Domain.Core/ProductCategoryAgg/Entities/ProductCategory.cs
@@ -18,7 +18,7 @@ namespace SM.Domain.Core.ProductCategoryAgg.Entities
         public string Name { get; private set; }
         public string Description { get; private set; }
         public DateTime CreationDate { get; private set; }
-        public int PictureId { get; private set; }
+        public long PictureId { get; private set; }
         public string MetaDescription { get; private set; }
         public string Slug { get; private set; }
         public string KeyWords { get; private set; }
